requirements.txt
openai
chromadb
python-dotenv
pydantic
streamlit
llama-index
llama-index[openai]
agno